+++ Compiler/script/cc_compiledscript.h
@@ -1,6 +1,7 @@
 #ifndef __CC_COMPILEDSCRIPT_H
 #define __CC_COMPILEDSCRIPT_H
 
+#include <cstdint>
 #include <string>
 #include "script/cc_script.h"       // ccScript
 #include "cs_parser_common.h"   // macro definitions
@@ -8,12 +9,12 @@
 
 
 struct ccCompiledScript: public ccScript {
-    long codeallocated;
+    int32_t codeallocated;
     char*functions[MAX_FUNCTIONS];
-    long funccodeoffs[MAX_FUNCTIONS];
+    int32_t funccodeoffs[MAX_FUNCTIONS];
     short funcnumparams[MAX_FUNCTIONS];
-    long numfunctions;
-    long cur_sp;
+    int32_t numfunctions;
+    int32_t cur_sp;
     int  next_line;  // line number of next code
     int  ax_val_type;  // type of value in AX
     int  ax_val_scope;  // SYM_LOCALVAR or SYM_GLOBALAVR
@@ -27,7 +28,7 @@ struct ccCompiledScript: public ccScript {
     void fixup_previous(char);
     int  add_new_function(const char*, int *idx);
     int  add_new_import(const char*);
-    int  add_new_export(const char*,int,long, int);
+    int  add_new_export(const char*, int, int32_t, int);
     void write_code(int32_t);
     void set_line_number(int nlum) { next_line=nlum; }
     void flush_line_numbers();
